homework-12/src/main/java/ru/otus/spring/homework12/services/GenreServiceImpl.java
package ru.otus.spring.homework12.services;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import ru.otus.spring.homework12.repositories.GenreRepository;
import ru.otus.spring.homework12.models.Genre;

import java.util.List;

@RequiredArgsConstructor
@Service
public class GenreServiceImpl implements GenreService {
    private final GenreRepository genreRepository;

    @Override
    public List<Genre> findAll() {
        return genreRepository.findAll();
    }
}
